refactor(#98): auth 관련 api 분리
